# New AR control field heuristics.