--- a/lib/jxl/encode_test.cc
+++ b/lib/jxl/encode_test.cc
@@ -259,7 +259,7 @@ TEST(EncodeTest, EncoderResetTest) {
   JxlEncoderPtr enc = JxlEncoderMake(nullptr);
   EXPECT_NE(nullptr, enc.get());
   VerifyFrameEncoding(50, 200, enc.get(),
-                      JxlEncoderFrameSettingsCreate(enc.get(), nullptr), 4550,
+                      JxlEncoderFrameSettingsCreate(enc.get(), nullptr), 4570,
                       false);
   // Encoder should become reusable for a new image from scratch after using
   // reset.