--- a/lib/jxl_lists.bzl
+++ b/lib/jxl_lists.bzl
@@ -317,8 +317,6 @@ libjxl_enc_sources = [
     "jxl/enc_ans.cc",
     "jxl/enc_ans.h",
     "jxl/enc_ans_params.h",
-    "jxl/enc_ar_control_field.cc",
-    "jxl/enc_ar_control_field.h",
     "jxl/enc_aux_out.cc",
     "jxl/enc_aux_out.h",
     "jxl/enc_bit_writer.cc",