… lib/jxl/enc_ar_control_field.cc
… lib/jxl/enc_ar_control_field.h
… lib/jxl/enc_heuristics.cc