The new algorithm bases the EPF strength decision on the actual reconstructed images filtered with the given EPF strength.

Benchmark results:
```
Encoding      kPixels    Bytes          BPP  E MP/s  D MP/s     Max norm  SSIMULACRA2   PSNR        pnorm       BPP*pnorm   QABPP   Bugs
----------------------------------------------------------------------------------------------------------------------------------------
BEFORE:
jxl:d0.1        20795 17960414    6.9092543   5.755  43.438   0.35676469  95.37045907  55.92   0.11743201  0.811367620746   6.909      0
jxl:d0.25       20795 11211481    4.3129837   6.651  47.980   0.47188711  93.95812476  51.27   0.19471022  0.839782033058   4.323      0
jxl:d0.5        20795  7343631    2.8250470   7.426  52.544   0.77713256  91.44634906  47.21   0.33192928  0.937715835158   2.911      0
jxl:d1          20795  4677692    1.7994777   8.299  54.165   1.34254171  86.90547291  43.27   0.56234123  1.011920474406   2.522      0
jxl:d2          20795  2886906    1.1105739   8.834  54.788   2.41345744  78.64070212  39.51   0.96381616  1.070389106583   2.822      0
jxl:d4          20795  1697421    0.6529868   8.108  42.320   4.09606455  65.00253202  36.14   1.60465567  1.047818979599   2.800      0
jxl:d8          20795  1000234    0.3847835   8.244  40.996   6.58003645  45.10259245  33.21   2.54751704  0.980242539117   2.576      0
Aggregate:      20795  4437904    1.7072329   7.546  47.724   1.41620256  77.25130756  43.13   0.55786685  0.952408666679   3.322      0
AFTER:
jxl:d0.1        20795 17960541    6.9093032   5.407  42.890   0.35676469  95.37045907  55.92   0.11743201  0.811373358013   6.909      0
jxl:d0.25       20795 11211603    4.3130307   6.196  47.317   0.47188711  93.95812476  51.27   0.19471022  0.839791171316   4.323      0
jxl:d0.5        20795  7343761    2.8250970   7.016  51.698   0.77713256  91.44634906  47.21   0.33192928  0.937732434992   2.911      0
jxl:d1          20795  4663158    1.7938865   5.824  52.251   1.34223458  86.92083387  43.31   0.56028046  1.005079568151   2.512      0
jxl:d2          20795  2883258    1.1091706   6.066  52.041   2.40615718  78.70689283  39.63   0.95343141  1.057518069706   2.807      0
jxl:d4          20795  1720412    0.6618313   5.059  40.640   4.02415306  65.76256057  36.37   1.56497555  1.035749795869   2.791      0
jxl:d8          20795  1031404    0.3967744   5.181  41.609   6.47993724  47.18952937  33.48   2.47074579  0.980328683551   2.621      0
Aggregate:      20795  4463209    1.7169674   5.788  46.672   1.40887758  77.89265621  43.25   0.55230474  0.948289250653   3.324      0
```

Images before:
https://storage.googleapis.com/jxl-quality/eval/szabadka/jyrki_split_2024-03-23/26e522d7/base/index.jxl_d2.html

Images after:
https://storage.googleapis.com/jxl-quality/eval/szabadka/jyrki_split_2024-03-23/ddd30717/base/index.jxl_d2.html