@szabadka 2024-04-24
Maybe I was using too many threads.

Here is a benchmark with --num_threads=8

before:

```
Encoding      kPixels    Bytes          BPP  E MP/s  D MP/s     Max norm  SSIMULACRA2   PSNR        pnorm       BPP*pnorm   QABPP   Bugs
----------------------------------------------------------------------------------------------------------------------------------------
jxl:d0.1        20795 17960414    6.9092543   5.667  43.639   0.35676469  95.37045907  55.92   0.11743201  0.811367620746   6.909      0
jxl:d0.25       20795 11211481    4.3129837   6.444  48.834   0.47188711  93.95812476  51.27   0.19471022  0.839782033058   4.323      0
jxl:d0.5        20795  7343631    2.8250470   7.257  53.099   0.77713256  91.44634906  47.21   0.33192928  0.937715835158   2.911      0
jxl:d1          20795  4677692    1.7994777   8.120  53.766   1.34254171  86.90547291  43.27   0.56234123  1.011920474406   2.522      0
jxl:d2          20795  2886906    1.1105739   8.623  53.864   2.41345744  78.64070212  39.51   0.96381616  1.070389106583   2.822      0
jxl:d4          20795  1697421    0.6529868   7.798  41.439   4.09606455  65.00253202  36.14   1.60465567  1.047818979599   2.800      0
jxl:d8          20795  1000234    0.3847835   7.994  40.631   6.58003645  45.10259245  33.21   2.54751704  0.980242539117   2.576      0
Aggregate:      20795  4437904    1.7072329   7.348  47.577   1.41620256  77.25130756  43.13   0.55786685  0.952408666679   3.322      0
```

after:
```
Encoding      kPixels    Bytes          BPP  E MP/s  D MP/s     Max norm  SSIMULACRA2   PSNR        pnorm       BPP*pnorm   QABPP   Bugs
----------------------------------------------------------------------------------------------------------------------------------------
jxl:d0.1        20795 17960541    6.9093032   5.017  39.054   0.35676469  95.37045907  55.92   0.11743201  0.811373358013   6.909      0
jxl:d0.25       20795 11211603    4.3130307   5.695  43.559   0.47188711  93.95812476  51.27   0.19471022  0.839791171316   4.323      0
jxl:d0.5        20795  7343761    2.8250970   6.435  46.917   0.77713256  91.44634906  47.21   0.33192928  0.937732434992   2.911      0
jxl:d1          20795  4661156    1.7931164   5.278  46.073   1.34183133  86.91182193  43.35   0.55814469  1.000818382521   2.514      0
jxl:d2          20795  2879199    1.1076091   5.412  46.431   2.41376322  78.59705840  39.64   0.95394965  1.056603315771   2.810      0
jxl:d4          20795  1717143    0.6605737   4.583  34.284   4.04711626  65.56610271  36.38   1.56647308  1.034770965746   2.802      0
jxl:d8          20795  1026482    0.3948809   4.674  35.303   6.49569992  46.74880752  33.48   2.47294840  0.976520194289   2.614      0
Aggregate:      20795  4457778    1.7148780   5.267  41.349   1.41108829  77.73841160  43.25   0.55219205  0.946942011366   3.326      0
```

It is still slower but the difference is less. Maybe we use filtering more often now.