@jyrkialakuijala 2024-04-24
Let's find a way not to use filtering more often