@jonsneyers 2024-04-24
While according to the metrics it seems to be an overall improvement in quality/bpp, this seems to be a substantial encoder slowdown too (going from 8 Mpx/s to 5 Mpx/s). Can this slowdown be mitigated?

How are other efforts than the default effort affected?