@szabadka 2024-04-24
Added a 0.93 factor bias for epf strength 0. Now single thread decoding speed is about the same with 8 threads it is about 2% slower.