@szabadka 2024-04-24
> While according to the metrics it seems to be an overall improvement in quality/bpp, this seems to be a substantial encoder slowdown too (going from 8 Mpx/s to 5 Mpx/s). Can this slowdown be mitigated?
> 
> How are other efforts than the default effort affected?

The encoder slowdown is expected, since we decode the image 3 times inside the encoder now. Since decode is about 10x faster than encode, this adds about 30% to the encode time.

Only efforts 6 and above are affected. Benchmark on effort 6:

```
Encoding       kPixels    Bytes          BPP  E MP/s  D MP/s     Max norm  SSIMULACRA2   PSNR        pnorm       BPP*pnorm   QABPP   Bugs
-----------------------------------------------------------------------------------------------------------------------------------------
BEFORE:
jxl:d0.1:6       20795 17859284    6.8703503   7.291  43.326   0.35909517  95.34837403  55.69   0.11899012  0.817503785154   6.872      0
jxl:d0.25:6      20795 11169872    4.2969770   8.576  47.963   0.48839340  93.88022760  51.06   0.19726629  0.847648706539   4.318      0
jxl:d0.5:6       20795  7279888    2.8005255  10.268  53.796   0.78525682  91.27818777  47.02   0.33551277  0.939612059770   2.887      0
jxl:d1:6         20795  4610232    1.7735262  11.908  53.817   1.37545408  86.55469305  43.11   0.57261271  1.015543669730   2.537      0
jxl:d2:6         20795  2832017    1.0894585  13.203  54.800   2.45583204  78.27084045  39.38   0.98143039  1.069227674344   2.806      0
jxl:d4:6         20795  1685543    0.6484174  14.146  43.978   4.09820795  64.81258457  36.09   1.61032213  1.044160913279   2.762      0
jxl:d8:6         20795   997515    0.3837375  14.713  43.939   6.58636036  44.85540890  33.16   2.55613817  0.980886135250   2.563      0
Aggregate:       20795  4399054    1.6922873  11.119  48.564   1.43543055  77.02999816  43.00   0.56428516  0.954932613117   3.306      0
AFTER:
jxl:d0.1:6       20795 17859412    6.8703995   6.837  42.757   0.35909517  95.34837403  55.69   0.11899012  0.817509644318   6.872      0
jxl:d0.25:6      20795 11170005    4.2970282   8.008  47.479   0.48839340  93.88022760  51.06   0.19726629  0.847658799517   4.318      0
jxl:d0.5:6       20795  7280015    2.8005744   9.501  51.800   0.78525682  91.27818777  47.02   0.33551277  0.939628451606   2.887      0
jxl:d1:6         20795  4595626    1.7679074   7.529  52.554   1.37131993  86.56306527  43.15   0.57029132  1.008222244600   2.520      0
jxl:d2:6         20795  2827975    1.0879036   7.786  53.118   2.45176829  78.33221870  39.51   0.97094780  1.056297572305   2.797      0
jxl:d4:6         20795  1708456    0.6572319   6.966  41.195   4.02716225  65.59488495  36.31   1.57116057  1.032616846768   2.750      0
jxl:d8:6         20795  1028977    0.3958408   7.171  43.237   6.44043500  46.99756946  33.43   2.47871579  0.981176721564   2.584      0
Aggregate:       20795  4424257    1.7019829   7.643  47.212   1.42632225  77.68787923  43.11   0.55865627  0.950823427558   3.303      0
```